modularize the browser tests
